pyproject.toml
[tool.poetry]
name = "golem"
version = "0.1.0"
description = ""
authors = ["Peder Landsverk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
numpy = "^1.26.3"
pywinauto = "^0.6.8"

[tool.poetry.scripts]
golem = "golem.entrypoint:entrypoint"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
